pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "yt-dlp",
    "flask",
    "openai>=1.0.0",
]
